Improved x86 64 bit support. Improved the indentation of generated code. Improved typing. Replaced some dirtiness with some gcc extensions. Validated at CRI. I'm able to parse and display some C code with PIPS in 64 bits at least, even there are still so many warning.
git-svn-id: http://svn.cri.ensmp.fr/svn/newgen/trunk@899 80f2d23d-f6f1-0310-8ea7-b0e6515bd809

+++ src/genC/list.c
@@ -95,14 +95,14 @@ int gen_eq(void * obj1, void * obj2)
     return obj1 == obj2;
 }
 
-int gen_length(list cp)
+size_t gen_length(list cp)
 {
   int i ;
   for( i=0; cp != NIL ; cp = cp->cdr, i++);
   return i;
 }
 
-int list_own_allocated_memory(list l)
+size_t list_own_allocated_memory(list l)
 {
     return gen_length(l)*sizeof(cons);
 }
